requirements.txt
streamlit
langchain
openai
pypdf2
python-docx
reportlab
pandas
python-dotenv
